+++ Modules/Intent.Modules.Blazor/Api/SelectModelStereotypeExtensions.cs
@@ -1,0 +1,75 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using Intent.Metadata.Models;
+using Intent.Modelers.UI.Core.Api;
+using Intent.Modules.Common;
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.ModuleBuilder.Templates.Api.ApiElementModelExtensions", Version = "1.0")]
+
+namespace Intent.Blazor.Api
+{
+    public static class SelectModelStereotypeExtensions
+    {
+        public static Secured GetSecured(this SelectModel model)
+        {
+            var stereotype = model.GetStereotype(Secured.DefinitionId);
+            return stereotype != null ? new Secured(stereotype) : null;
+        }
+
+        public static IReadOnlyCollection<Secured> GetSecureds(this SelectModel model)
+        {
+            var stereotypes = model
+                .GetStereotypes(Secured.DefinitionId)
+                .Select(stereotype => new Secured(stereotype))
+                .ToArray();
+
+            return stereotypes;
+        }
+
+
+        public static bool HasSecured(this SelectModel model)
+        {
+            return model.HasStereotype(Secured.DefinitionId);
+        }
+
+        public static bool TryGetSecured(this SelectModel model, out Secured stereotype)
+        {
+            if (!HasSecured(model))
+            {
+                stereotype = null;
+                return false;
+            }
+
+            stereotype = new Secured(model.GetStereotype(Secured.DefinitionId));
+            return true;
+        }
+
+        public class Secured
+        {
+            private IStereotype _stereotype;
+            public const string DefinitionId = "012f5173-6419-4006-a9a8-ab5c20b8a42e";
+
+            public Secured(IStereotype stereotype)
+            {
+                _stereotype = stereotype;
+            }
+
+            public string Name => _stereotype.Name;
+
+            public string Roles()
+            {
+                return _stereotype.GetProperty<string>("Roles");
+            }
+
+            public string Policy()
+            {
+                return _stereotype.GetProperty<string>("Policy");
+            }
+
+        }
+
+    }
+}