+++ Modules/Intent.Modules.Blazor/NugetPackages.cs
@@ -12,6 +12,7 @@ namespace Intent.Modules.Blazor
     public class NugetPackages : INugetPackages
     {
         public const string MicrosoftAspNetCoreComponentsWebAssemblyPackageName = "Microsoft.AspNetCore.Components.WebAssembly";
+        public const string MicrosoftAspNetCoreComponentsWebAssemblyAuthenticationPackageName = "Microsoft.AspNetCore.Components.WebAssembly.Authentication";
         public const string MicrosoftAspNetCoreComponentsWebAssemblyServerPackageName = "Microsoft.AspNetCore.Components.WebAssembly.Server";
 
         public void RegisterPackages()
@@ -46,6 +47,27 @@ namespace Intent.Modules.Blazor
                         _ => throw new Exception($"Unsupported Framework `{framework.Major}` for NuGet package '{MicrosoftAspNetCoreComponentsWebAssemblyPackageName}'"),
                     }
                 );
+            NugetRegistry.Register(MicrosoftAspNetCoreComponentsWebAssemblyAuthenticationPackageName,
+                (framework) => (framework.Major, framework.Minor) switch
+                    {
+                        ( >= 9, >= 0) => new PackageVersion("9.0.5")
+                            .WithNugetDependency("Microsoft.AspNetCore.Components.Authorization", "9.0.5")
+                            .WithNugetDependency("Microsoft.AspNetCore.Components.Web", "9.0.5"),
+                        ( >= 8, >= 0) => new PackageVersion("8.0.16")
+                            .WithNugetDependency("Microsoft.AspNetCore.Components.Authorization", "8.0.16")
+                            .WithNugetDependency("Microsoft.AspNetCore.Components.Web", "8.0.16"),
+                        ( >= 7, >= 0) => new PackageVersion("7.0.20")
+                            .WithNugetDependency("Microsoft.AspNetCore.Components.Authorization", "7.0.20")
+                            .WithNugetDependency("Microsoft.AspNetCore.Components.Web", "7.0.20"),
+                        ( >= 6, >= 0) => new PackageVersion("6.0.36")
+                            .WithNugetDependency("Microsoft.AspNetCore.Components.Authorization", "6.0.36")
+                            .WithNugetDependency("Microsoft.AspNetCore.Components.Web", "6.0.36"),
+                        ( >= 2, >= 1) => new PackageVersion("3.2.1")
+                            .WithNugetDependency("Microsoft.AspNetCore.Components.Authorization", "3.1.6")
+                            .WithNugetDependency("Microsoft.AspNetCore.Components.Web", "3.1.6"),
+                        _ => throw new Exception($"Unsupported Framework `{framework.Major}` for NuGet package '{MicrosoftAspNetCoreComponentsWebAssemblyAuthenticationPackageName}'"),
+                    }
+                );
             NugetRegistry.Register(MicrosoftAspNetCoreComponentsWebAssemblyServerPackageName,
                 (framework) => (framework.Major, framework.Minor) switch
                     {
@@ -59,6 +81,8 @@ namespace Intent.Modules.Blazor
         }
 
         public static NugetPackageInfo MicrosoftAspNetCoreComponentsWebAssembly(IOutputTarget outputTarget) => NugetRegistry.GetVersion(MicrosoftAspNetCoreComponentsWebAssemblyPackageName, outputTarget.GetMaxNetAppVersion());
+
+        public static NugetPackageInfo MicrosoftAspNetCoreComponentsWebAssemblyAuthentication(IOutputTarget outputTarget) => NugetRegistry.GetVersion(MicrosoftAspNetCoreComponentsWebAssemblyAuthenticationPackageName, outputTarget.GetMaxNetAppVersion());
 
         public static NugetPackageInfo MicrosoftAspNetCoreComponentsWebAssemblyServer(IOutputTarget outputTarget) => NugetRegistry.GetVersion(MicrosoftAspNetCoreComponentsWebAssemblyServerPackageName, outputTarget.GetMaxNetAppVersion());
     }