+++ Modules/Intent.Modules.Blazor.HttpClients.Dtos.FluentValidation/release-notes.md
@@ -1,5 +1,6 @@
 ### Version 1.2.7
 
+- Improvement: Updated referenced packages versions
 - Fixed: Under certain circumstances the Software Factory would fail with a "More than one instance of `<templateId>` was found with `<modelId>`" error.
 
 ### Version 1.2.6